+++ prettier.config.js
@@ -1,0 +1,1 @@
+module.exports = require('@meludi/eslint-config-prettier/prettier.config');