+++ packages/eslint-config-base/CHANGELOG.md
@@ -5,7 +5,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # 0.1.0 (2023-02-24)
 
-
 ### Features
 
-* **eslint-config-base:** create package ([879026c](https://github.com/meludi/eslint-config/commit/879026c172df3dc10753ceb91a7424f87c08b6e3))
+- **eslint-config-base:** create package ([879026c](https://github.com/meludi/eslint-config/commit/879026c172df3dc10753ceb91a7424f87c08b6e3))