+++ packages/eslint-config-prettier/CHANGELOG.md
@@ -5,7 +5,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # 0.1.0 (2023-02-24)
 
-
 ### Features
 
-* **eslint-config-prettier:** create package ([dbddba2](https://github.com/meludi/eslint-config/commit/dbddba235fbe4a6ff650dd97f89fe618e5c7ee43))
+- **eslint-config-prettier:** create package ([dbddba2](https://github.com/meludi/eslint-config/commit/dbddba235fbe4a6ff650dd97f89fe618e5c7ee43))